Bulk RNA_Seq Data Analysis
Bulk RNA-Seq is used to compare the gene expression patterns between samples. This is a bulk RNA-Seq analysis of the saccharomyces cerevisiae (mRNA) dataset, which is 103.4 MB, taken from NCBI-SRA SRR5924196. The data then underwent preprocessing for quality control and adapter trimming for the removal of low-quality reads and adapter contamination. Further, after generating the count matrix, the preprocessed data is used for downstream analysis using the DESeq package of R, and the expression pattern is visualized through volcano plots, heatmaps, and various others.