ScRNA Workflow
This is the ScRNA workflow of the NSCLC dataset from 10X Genomics. To study the gene expression from a particular cell and to analyze the effect of this particular cell on the neighboring cells within the tissue sample. For this study, the Seurat library (part of a Bioconductor package) has been used to study the complete gene expression profile and also the tidyverse library for data manipulation.