Spatial transcriptomics-mouse brain dataset
Identifying the brain cells' activity from the mouse dataset through a molecular profiling technique called spatial transcriptomics. One of our main targets is to resolve the cell heterogeneity, keeping track of where this activity occurs in the tissue sample. The dataset (stxBrain) is taken from 10X genomics and is also a part of the Seurat package.
Packages used for this pipeline include Seurat, SeuratData, and ggplot2